Assignments/Team Lead/Assignment 1/requirements.txt
Flask~=2.2.2
email-validator~=1.3.0
requests~=2.28.1
python-avatars~=1.3.1
python-dateutil~=2.8.2
